fix: force fp32 in PixelDiffusionInferencer to avoid mixed-precision mat errors
Same root cause as the gapfill fix: resolve_dtype defaults to fp16 on CUDA,
producing unet512 with fp16 weights while the input x is created with self.dtype.
Hardcode float32 for build_inference_components to keep dtypes consistent.

+++ morphe/core/pixel_diffusion/inferencer.py
@@ -76,7 +76,7 @@ class PixelDiffusionInferencer:
         adapter, unet512, noise_scheduler, resolved_device, resolved_dtype = (
             artifact.build_inference_components(
                 device=device,
-                dtype=dtype,
+                dtype=torch.float32,
                 scheduler_num_inference_steps=scheduler_num_inference_steps
             )
         )
